% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funnyDatasets.R
\docType{data}
\name{chopstickEffect}
\alias{chopstickEffect}
\title{Chopstick Effectiveness}
\format{A dataa frame with 186 rows, and 3 variables:
\itemize{
 \item Food.Pinching.Efficiency.
     The number of peanuts picked and placed in cup numbers (PPPC).
 \item Individual.
 \item Chopstick.Length. Lenght in millimiters.
}}
\source{
\url{http://blog.yhathq.com/static/misc/data/chopstick-effectiveness.csv}
}
\usage{
chopstickEffect
}
\description{
The dataset is part of the paper "An investigation for
detemining the optimum lenght of chopsticks",
published on the Journal Applied Ergonomics, in 1991.
The authors try to determine the effects
of the length of the chopsticks on the food-serving performance
of adults and children.
The original paper can be found at
\url{"http://www.sciencedirect.com/science/article/pii/000368709190082S"};
doi:10.1016/0003-6870(91)90082-S
}
\references{
\url{"http://www.sciencedirect.com/science/article/pii/000368709190082S"}
}
\keyword{datasets}

