% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funnyDatasets.R
\docType{data}
\name{marijuanaStPrice}
\alias{marijuanaStPrice}
\title{Marijuana street price}
\format{A data frame with 22899 rows and 8 variables:
\itemize{
 \item State. US states.
 \item HighQ. The average price for the high quality product,
     in US dollars.
 \item HighQN. Number of observations, high quality product.
 \item MedQ. The average price for the medium quality product,
     in US dollars.
 \item MedQN. Number of observations, medium quality product.
 \item LowQ. The average price for the low quality product,
     in US dollars.
 \item LowQN. Number of observations, low quality product.
 \item date. Date of the year, in the format "Year-month-day".
}}
\source{
\url{http://blog.yhathq.com/static/misc/data/marijuana-street-price-clean.csv}
}
\usage{
marijuanaStPrice
}
\description{
A dataset containing the street prices of marijuana in the US.
Information about the data collection process can be found at
\url{https://github.com/frankbi/price-of-weed}
}
\keyword{datasets}

