% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funnyDatasets.R
\docType{data}
\name{pigeonsRacing}
\alias{pigeonsRacing}
\title{Pigeon Racing}
\format{
\itemize{
 \item Pos. The position at the arrival.
 \item Breeder. Name/nickname of the breader.
 \item Pigeon. Pigeon ID
 \item Name. Name of the pigeon, if any
 \item Color. Color code
 \item Sex. Two levels: "H" (hen, female), and "C" (cock, male)
 \item Ent.
 \item Arrival. Overall racing time
 \item Speed. Average speed
 \item To Win. Time difference from the winner.
 \item Eligible. Binary variable representing
     whether the performace is a valid ARPU competition.
}}
\source{
{http://blog.yhathq.com/static/misc/data/pigeon-racing.csv}
}
\usage{
pigeonsRacing
}
\description{
A dataset containig a selection of records of pigeon racing from the
American Racing Pigeon Union. More info are available
at \url{http://www.pigeon.org/}
}
\keyword{datasets}

