% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funnyDatasets.R
\docType{data}
\name{spanishSilver}
\alias{spanishSilver}
\title{Spanish silver production between 1720-1800}
\format{A data frame with 81 rows and 4 variables:
\itemize{
 \item year. The year
 \item situados paid. The number of paied miners
 \item silver minted. The amount of silver minted
 \item was american revolution. Binary variable: 0 is FALSE, 1 is TRUE.
}}
\source{
\url{http://blog.yhathq.com/static/misc/data/spanish-silver.csv}
}
\usage{
spanishSilver
}
\description{
An historical datset that keeps track of the silver extraction
during the colonial era.
}
\keyword{datasets}

