% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funnyDatasets.R
\docType{data}
\name{speedStackingTimes}
\alias{speedStackingTimes}
\title{Speed Stacking Times}
\format{A data frame with 130 rows and 6 variables:
\itemize{
 \item Competitor. Participant ID in the format <First Name> <Last Name>.
 \item Division. Age and Gender division. M for "Male", and F for "Female".
     "U" for "under". "Collegiate" (19-24),
     "Masters 1" (25-34, "Masters 2" (35-44), "Masters 3" (45-59).
     Full details are available at
     "http://www.thewssa.com/rules/events_and_age_divisions/").
 \item Stack. Type of event.
     Four categories are represented in this dataset
     out of the official six. The full list,
     with a detailed description is available
     at http://www.thewssa.com/competitive_stacks/.
 \item Time. Time, in seconds.
 \item Type. It is a sanctioned tournament.
 \item Location. Name of the competition.
}}
\source{
\url{http://blog.yhathq.com/static/misc/data/world-sport-stacking.csv}
}
\usage{
speedStackingTimes
}
\description{
A dataset containing a selection of records from the World Sport Stacking Association (WSSA)
}
\keyword{datasets}

